+++ deployment/entry_point_celery_flower.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-celery -A "iometrics_alerta.plugins.bgtasks.celery" flower --port="${CELERY_FLOWER_PORT}"
+celery -A "datadope_alerta.bgtasks.celery" flower --port="${CELERY_FLOWER_PORT}"