+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 2.0.0
+
+* Change main package name to datadope_alerta
+* Blackouts management plugin
+
+# 1.1.0
+
+* API contexts to read alerters information and to process an alert asynchronously
+* Refactor: bgtasks moved to package first level
+
+# 1.0.1
+
+* Catch unique exception when creating alerter status.
+
 # 1.0.0
 
 Backend: Deduplicate using attribute `deduplication`